% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-extra-functions.R
\name{add_entry}
\alias{add_entry}
\title{Adding an entry to or creating a timesheet}
\usage{
add_entry(start, end, work, timesheet = NULL)
}
\arguments{
\item{start}{Start of work.}

\item{end}{End of work.}

\item{work}{What was done?}

\item{timesheet}{A correctly named data.table.}
}
\value{
A timesheet.
}
\description{
Adding an entry to or creating a timesheet
}
\examples{
timesheet <- add_entry("2021-12-04 12:30:00", "2021-12-04 13:30:00", "Eating in office hours")
timesheet <- add_entry("2021-12-04 13:30:00", "2021-12-04 16:30:00", "Powernap", timesheet)
}
