% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting_record.R
\name{meeting_record}
\alias{meeting_record}
\title{Create a summary data.frame of meetings within a date interval}
\usage{
meeting_record(
  name,
  start_date,
  end_date,
  remove_text = " and Marcel Ramos",
  include_cancel = FALSE
)
}
\arguments{
\item{name}{character(1) The Google calendar name}

\item{start_date}{character(1) The start date to restrict the entries in
'YYYY-MM-DD' format.}

\item{end_date}{character(1) The end date to restrict the entries
\code{data.frame}.}

\item{include_cancel}{logical(1) Whether or not to include entried beginning
with "Canceled" (default: FALSE)}
}
\value{
A \code{data.frame} of meeting times and student information
}
\description{
Query a particular calendar name to extract the event entries. The associated
metadata will be included, e.g., first and last names, program, class,
EMPLID.
}
\examples{

if (interactive()) {
  meeting_record(
    name = "Quantitative",
    start_date = "2022-09-25",
    end_date = "2022-10-08"
  )
}

}
