% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-extra-functions.R
\name{read_timesheet}
\alias{read_timesheet}
\title{Function for reading in a timesheet. (Save timesheet as csv.)}
\usage{
read_timesheet(path)
}
\arguments{
\item{path}{Path to timesheet.}
}
\value{
timesheet data.table
}
\description{
Function for reading in a timesheet. (Save timesheet as csv.)
}
\examples{
\dontrun{
timesheet <- read_timesheet("timesheet.csv")
utils::write.csv(timesheet, "timesheet.csv")
}
}
