% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-extra-functions.R
\name{time_add}
\alias{time_add}
\title{Function for adding unusual time formats.}
\usage{
time_add(ls_time)
}
\arguments{
\item{ls_time}{List of times in "4h 45min" format.}
}
\value{
A sum in "4h 45min" format.
}
\description{
Function for adding unusual time formats.
}
\examples{
timesheet <- add_entry("2021-11-04 12:30:00", "2021-12-04 13:30:00", "Eating in office hours")
timesheet <- add_entry("2021-12-04 13:30:00", "2021-12-04 16:30:00", "Powernap", timesheet)
lapply(11:12, time_month, timesheet = timesheet, work_left = TRUE) \%>\% time_add()
}
