% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-aggregation.R
\name{time_month}
\alias{time_month}
\title{Sum up the working hours of month i.}
\usage{
time_month(timesheet, month, work_left = FALSE, hours_per_month = 39)
}
\arguments{
\item{timesheet}{A correctly named data.table.}

\item{month}{Am integer between 1 and 12.}

\item{work_left}{Calculate remaining hours instead.}

\item{hours_per_month}{How many hours are worked per month.}
}
\value{
A string with time in hours.
}
\description{
Sum up the working hours of month i.
}
\examples{
timesheet <- add_entry("2021-11-04 12:30:00", "2021-12-04 13:30:00", "Eating in office hours")
timesheet <- add_entry("2021-12-04 13:30:00", "2021-12-04 16:30:00", "Powernap", timesheet)
time_month(timesheet, month = 11)
time_month(timesheet, month = 12)
time_month(timesheet, month = 12, work_left = TRUE)
}
