% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-aggregation.R
\name{time_total}
\alias{time_total}
\title{Sum up all working hours.}
\usage{
time_total(timesheet)
}
\arguments{
\item{timesheet}{A correctly named data.table.}
}
\value{
A string with time in hours.
}
\description{
Sum up all working hours.
}
\examples{
timesheet <- add_entry("2021-11-04 12:30:00", "2021-11-04 13:30:00", "Eating in office hours")
timesheet <- add_entry("2021-12-04 13:30:00", "2021-12-04 16:30:00", "Powernap", timesheet)
time_total(timesheet)
}
