% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheet-package.R
\docType{package}
\name{timesheet-package}
\alias{timesheet}
\alias{timesheet-package}
\title{timesheet: Keep Track Of Your Working Hours}
\description{
Functions for maintaining and evaluating timesheets.
}
\author{
\strong{Maintainer}: Max Last \email{first.last@example.com}

}
\keyword{internal}
