% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesheet.R
\name{timesheet}
\alias{timesheet}
\title{Tally of hours per day within a given period from Google calendar}
\usage{
timesheet(
  cal_name = "Quantitative",
  start_date,
  end_date,
  include_cancel = FALSE,
  token = getOption("timesheet.token"),
  maxResults = 2000L
)
}
\description{
Provides the user a small table of total hours per day, within a given
Google Calendar name.
}
\examples{

timesheet("Quantitative", "2022-08-28", "2022-09-10")

}
